+++ src/widgets/containers/center.rs
@@ -15,6 +15,7 @@ impl Default for Center {
 
 impl Center {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {