+++ src/widgets/containers/center_middle.rs
@@ -15,6 +15,7 @@ impl Default for CenterMiddle {
 
 impl CenterMiddle {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {