+++ src/widgets/containers/horizontal.rs
@@ -16,6 +16,7 @@ impl Default for Horizontal {
 
 impl Horizontal {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {