+++ src/widgets/containers/horizontal_group.rs
@@ -16,6 +16,7 @@ impl Default for HorizontalGroup {
 
 impl HorizontalGroup {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {