+++ src/widgets/containers/middle.rs
@@ -15,6 +15,7 @@ impl Default for Middle {
 
 impl Middle {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {