+++ src/widgets/containers/right.rs
@@ -15,6 +15,7 @@ impl Default for Right {
 
 impl Right {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {