+++ src/widgets/containers/scrollable_container.rs
@@ -13,6 +13,7 @@ pub struct ScrollableContainer {
 
 impl ScrollableContainer {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {