+++ src/widgets/containers/vertical_group.rs
@@ -21,6 +21,7 @@ impl Default for VerticalGroup {
 
 impl VerticalGroup {
     crate::delegate_ident_methods!(inner);
+    crate::delegate_border_title_methods!(inner);
 
     pub fn new() -> Self {
         Self {